+++ astropy/modeling/models.py
@@ -15,7 +15,7 @@ from .polynomial import *
 from .functional_models import *
 from .powerlaws import *
 from .tabular import *
-from .blackbody import *
+from .blackbody import BlackBody1D
 
 from ..extern.six import string_types